Migrated QR codes to user service, from checkin service.
Moved controllers
Moved model
Moved services
Moved tests
Moved gateway routes
Moved documentation
