+++ services/checkin/docs/Checkin.md
@@ -76,31 +76,3 @@ Response format:
 	"hasPickedUpSwag": true
 }
 ```
-
-GET /checkin/qr/
-----------
-
-Get the string to be embedded in the current user's QR code. 
-The QR code string will contain information stored in the form of a URI.
-
-Response format:
-```
-{
-	"id": "github0000001",
-	"qrInfo": "hackillinois://info?name=FirstName&hasCheckedIn=true&hasPickedUpSwag=true"
-}
-```
-
-GET /checkin/qr/{id}/
-----------
-
-Get the string to be embedded in the specified user's QR code. 
-The QR code string will contain information stored in the form of a URI.
-
-Response format:
-```
-{
-	"id": "github0000001",
-	"qrInfo": "hackillinois://info?name=FirstName&hasCheckedIn=true&hasPickedUpSwag=true"
-}
-```
